# backend/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/admin/drives/{driveId}": {
            "get": {
                "tags": [
                    "admin-placement-drive-controller"
                ],
                "operationId": "getDriveById",
                "parameters": [
                    {
                        "name": "driveId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePlacementDriveResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "admin-placement-drive-controller"
                ],
                "operationId": "updateDrive",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "driveId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlacementDriveRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePlacementDriveResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/student/auth/register": {
            "post": {
                "tags": [
                    "student-auth-controller"
                ],
                "operationId": "register",
                "parameters": [
                    {
                        "name": "universityRegNo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/auth/login": {
            "post": {
                "tags": [
                    "student-auth-controller"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAuthResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/faculty/events": {
            "get": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "getEvents",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListEvent"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "createEvent",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Event"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseEvent"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/faculty/announcements": {
            "get": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "getAnnouncements",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListAnnouncement"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "faculty-announcement-controller"
                ],
                "operationId": "createAnnouncement",
                "parameters": [
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Announcement"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAnnouncement"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/student/drives": {
            "get": {
                "tags": [
                    "student-drive-controller"
                ],
                "operationId": "getVisibleDrives",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListPlacementDriveDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/student/drives/{driveId}": {
            "get": {
                "tags": [
                    "student-drive-controller"
                ],
                "operationId": "getDriveDetails",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "driveId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePlacementDriveDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/faculty/students/{id}": {
            "get": {
                "tags": [
                    "faculty-student-controller"
                ],
                "operationId": "getStudentProfile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "facultyEmail",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "faculty@dept.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseFacultyStudentDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/students/{studentId}": {
            "get": {
                "tags": [
                    "admin-student-controller"
                ],
                "operationId": "getStudentById",
                "parameters": [
                    {
                        "name": "studentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAdminStudentProfileDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "PlacementDriveRequestDTO": {
                "type": "object",
                "properties": {
                    "companyId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "ctcLpa": {
                        "type": "number",
                        "format": "double"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "UPCOMING",
                            "ONGOING",
                            "COMPLETED"
                        ]
                    }
                },
                "required": [
                    "companyId",
                    "ctcLpa",
                    "role",
                    "title"
                ]
            },
            "ApiResponsePlacementDriveResponseDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PlacementDriveResponseDTO"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PlacementDriveResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "companyId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "companyName": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "ctcLpa": {
                        "type": "number",
                        "format": "double"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "UPCOMING",
                            "ONGOING",
                            "COMPLETED"
                        ]
                    }
                }
            },
...
            "AuthRequestDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "ApiResponseAuthResponseDto": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AuthResponseDto"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AuthResponseDto": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
...
            "Department": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
            "Event": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "scheduledAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "locationOrLink": {
                        "type": "string"
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    }
                }
            },
            "ApiResponseEvent": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Event"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Announcement": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "scope": {
                        "type": "string",
                        "enum": [
                            "GLOBAL",
                            "DEPARTMENT"
                        ]
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/User"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "type": "string"
                    },
                    "universityRegNo": {
                        "type": "string"
                    },
                    "passwordHash": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "STUDENT",
                            "FACULTY",
                            "PLACEMENT_HEAD"
                        ]
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "ApiResponseAnnouncement": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Announcement"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseAdminStudentProfileDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AdminStudentProfileDTO"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListEvent": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Event"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListAnnouncement": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Announcement"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseListPlacementDriveDto": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PlacementDriveDto"
                        }
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PlacementDriveDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "ctcLpa": {
                        "type": "number",
                        "format": "double"
                    },
                    "status": {
                        "type": "string"
                    },
                    "companyName": {
                        "type": "string"
                    },
                    "companyIndustry": {
                        "type": "string"
                    },
                    "companyWebsite": {
                        "type": "string"
                    },
                    "isEligible": {
                        "type": "boolean"
                    },
                    "ineligibilityReason": {
                        "type": "string"
                    }
                }
            },
            "ApiResponsePlacementDriveDto": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PlacementDriveDto"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "FacultyStudentDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "rollNo": {
                        "type": "string"
                    },
                    "batch": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "verificationStatus": {
                        "type": "string"
                    },
                    "isLocked": {
                        "type": "boolean"
                    },
                    "isEligibleForPlacements": {
                        "type": "boolean"
                    },
                    "isPlaced": {
                        "type": "boolean"
                    },
                    "numberOfOffers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "highestPackageLpa": {
                        "type": "number",
                        "format": "double"
                    },
                    "cgpa": {
                        "type": "number",
                        "format": "double"
                    },
                    "standingArrears": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "historyOfArrears": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ApiResponseFacultyStudentDTO": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/FacultyStudentDTO"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}